+++ api/tests/unit/segments/test_unit_segments_views.py
@@ -27,7 +27,6 @@ from features.versioning.models import EnvironmentFeatureVersion
 from metadata.models import Metadata, MetadataModelField
 from projects.models import Project
 from segments.models import Condition, Segment, SegmentRule, WhitelistedSegment
-from segments.services import SegmentCloneService
 from tests.types import WithProjectPermissionsCallable
 from util.mappers import map_identity_to_identity_document
 
@@ -153,10 +152,9 @@ def test_create_segments_reaching_max_limit(project, client, settings):  # type:
     # Let's try to create a second segment
     res = client.post(url, data=json.dumps(data), content_type="application/json")
     assert res.status_code == status.HTTP_400_BAD_REQUEST
-    assert (
-        res.json()["project"]
-        == "The project has reached the maximum allowed segments limit."
-    )
+    assert res.json()["project"] == [
+        "The project has reached the maximum allowed segments limit."
+    ]
     assert project.segments.count() == 1
 
 
@@ -174,8 +172,8 @@ def test_segments_limit_ignores_old_segment_versions(
     project.save()
 
     # and create some older versions for the segment fixture
-    SegmentCloneService(segment).deep_clone()
-    assert Segment.objects.filter(version_of_id=segment.id).count() == 3
+    segment.clone(is_revision=True)
+    assert Segment.objects.filter(version_of=segment).count() == 2
     assert Segment.live_objects.count() == 1
 
     url = reverse("api-v1:projects:project-segments-list", args=[project.id])
@@ -202,9 +200,12 @@ def test_segments_limit_ignores_old_segment_versions(
     "client",
     [lazy_fixture("admin_master_api_key_client"), lazy_fixture("admin_client")],
 )
-def test_audit_log_created_when_segment_updated(project, client):  # type: ignore[no-untyped-def]
+def test_audit_log_created_when_segment_updated(
+    client: APIClient,
+    project: Project,
+    segment: Segment,
+) -> None:
     # Given
-    segment = Segment.objects.create(name="Test segment", project=project)
     url = reverse(
         "api-v1:projects:project-segments-detail",
         args=[project.id, segment.id],
@@ -221,12 +222,10 @@ def test_audit_log_created_when_segment_updated(project, client):  # type: ignor
     # Then
     assert response.status_code == status.HTTP_200_OK
 
-    assert (
-        AuditLog.objects.filter(
-            related_object_type=RelatedObjectType.SEGMENT.name
-        ).count()
-        == 1
-    )
+    assert AuditLog.objects.filter(
+        related_object_type=RelatedObjectType.SEGMENT.name,
+        log="Segment updated: New segment name",
+    ).exists()
 
 
 @pytest.mark.parametrize(
@@ -712,148 +711,6 @@ def test_update_segment_add_new_condition(
     assert nested_rule.conditions.last().value == new_condition_value
 
 
-def test_update_mismatched_rule_and_segment(
-    project: Project,
-    admin_client_new: APIClient,
-    segment: Segment,
-    segment_rule: SegmentRule,
-) -> None:
-    # Given
-    url = reverse(
-        "api-v1:projects:project-segments-detail", args=[project.id, segment.id]
-    )
-    false_segment = Segment.objects.create(name="False segment", project=project)
-    segment_rule.segment = false_segment
-    segment_rule.save()
-
-    nested_rule = SegmentRule.objects.create(
-        rule=segment_rule, type=SegmentRule.ANY_RULE
-    )
-    existing_condition = Condition.objects.create(
-        rule=nested_rule, property="foo", operator=EQUAL, value="bar"
-    )
-
-    new_condition_property = "foo2"
-    new_condition_value = "bar"
-    data = {
-        "name": segment.name,
-        "project": project.id,
-        "rules": [
-            {
-                "id": segment_rule.id,
-                "type": segment_rule.type,
-                "rules": [
-                    {
-                        "id": nested_rule.id,
-                        "type": nested_rule.type,
-                        "rules": [],
-                        "conditions": [
-                            # existing condition
-                            {
-                                "id": existing_condition.id,
-                                "property": existing_condition.property,
-                                "operator": existing_condition.operator,
-                                "value": existing_condition.value,
-                            },
-                            # new condition
-                            {
-                                "property": new_condition_property,
-                                "operator": EQUAL,
-                                "value": new_condition_value,
-                            },
-                        ],
-                    }
-                ],
-                "conditions": [],
-            }
-        ],
-    }
-
-    # When
-    response = admin_client_new.put(
-        url, data=json.dumps(data), content_type="application/json"
-    )
-
-    # Then
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json() == {"segment": "Mismatched segment is not allowed"}
-    segment_rule.refresh_from_db()
-    assert segment_rule.segment == false_segment
-
-
-def test_update_mismatched_condition_and_segment(
-    project: Project,
-    admin_client_new: APIClient,
-    segment: Segment,
-    segment_rule: SegmentRule,
-) -> None:
-    # Given
-    url = reverse(
-        "api-v1:projects:project-segments-detail", args=[project.id, segment.id]
-    )
-    false_segment = Segment.objects.create(name="False segment", project=project)
-    false_segment_rule = SegmentRule.objects.create(
-        segment=false_segment, type=SegmentRule.ALL_RULE
-    )
-    false_nested_rule = SegmentRule.objects.create(
-        rule=false_segment_rule, type=SegmentRule.ANY_RULE
-    )
-    nested_rule = SegmentRule.objects.create(
-        rule=segment_rule, type=SegmentRule.ANY_RULE
-    )
-
-    existing_condition = Condition.objects.create(
-        rule=false_nested_rule, property="foo", operator=EQUAL, value="bar"
-    )
-
-    new_condition_property = "foo2"
-    new_condition_value = "bar"
-    data = {
-        "name": segment.name,
-        "project": project.id,
-        "rules": [
-            {
-                "id": segment_rule.id,
-                "type": segment_rule.type,
-                "rules": [
-                    {
-                        "id": nested_rule.id,
-                        "type": nested_rule.type,
-                        "rules": [],
-                        "conditions": [
-                            # existing condition
-                            {
-                                "id": existing_condition.id,
-                                "property": existing_condition.property,
-                                "operator": existing_condition.operator,
-                                "value": existing_condition.value,
-                            },
-                            # new condition
-                            {
-                                "property": new_condition_property,
-                                "operator": EQUAL,
-                                "value": new_condition_value,
-                            },
-                        ],
-                    }
-                ],
-                "conditions": [],
-            }
-        ],
-    }
-
-    # When
-    response = admin_client_new.put(
-        url, data=json.dumps(data), content_type="application/json"
-    )
-
-    # Then
-    assert response.status_code == status.HTTP_400_BAD_REQUEST
-    assert response.json() == {"segment": "Mismatched segment is not allowed"}
-    existing_condition.refresh_from_db()
-    assert existing_condition._get_segment() != segment
-
-
 def test_update_segment_versioned_segment(
     project: Project,
     admin_client_new: APIClient,
@@ -870,10 +727,6 @@ def test_update_segment_versioned_segment(
     existing_condition = Condition.objects.create(
         rule=nested_rule, property="foo", operator=EQUAL, value="bar"
     )
-
-    # Before updating the segment confirm pre-existing version count which is
-    # automatically set by the fixture.
-    assert Segment.objects.filter(version_of=segment).count() == 2
 
     new_condition_property = "foo2"
     new_condition_value = "bar"
@@ -920,11 +773,11 @@ def test_update_segment_versioned_segment(
     assert response.status_code == status.HTTP_200_OK
 
     # Now verify that a new versioned segment has been set.
-    assert Segment.objects.filter(version_of=segment).count() == 3
+    assert Segment.objects.filter(version_of=segment).count() == 2
 
     # Now check the previously versioned segment to match former count of conditions.
 
-    versioned_segment = Segment.objects.filter(version_of=segment, version=2).first()
+    versioned_segment = Segment.objects.filter(version_of=segment, version=1).first()
     assert versioned_segment != segment
     assert versioned_segment.rules.count() == 1
     versioned_rule = versioned_segment.rules.first()
@@ -955,7 +808,7 @@ def test_update_segment_versioned_segment_with_thrown_exception(
         rule=nested_rule, property="foo", operator=EQUAL, value="bar"
     )
 
-    assert segment.version == 2 == Segment.objects.filter(version_of=segment).count()
+    assert segment.version == 1 == Segment.objects.filter(version_of=segment).count()
 
     new_condition_property = "foo2"
     new_condition_value = "bar"
@@ -1006,7 +859,7 @@ def test_update_segment_versioned_segment_with_thrown_exception(
     segment.refresh_from_db()
 
     # Now verify that the version of the segment has not been changed.
-    assert segment.version == 2 == Segment.objects.filter(version_of=segment).count()
+    assert segment.version == 1 == Segment.objects.filter(version_of=segment).count()
 
 
 @pytest.mark.parametrize(
@@ -1340,7 +1193,9 @@ def test_update_segment_obeys_max_conditions(
     # Then
     assert response.status_code == status.HTTP_400_BAD_REQUEST
     assert response.json() == {
-        "segment": "The segment has 11 conditions, which exceeds the maximum condition count of 10."
+        "segment": [
+            "The segment has 11 conditions, which exceeds the maximum condition count of 10."
+        ]
     }
 
     nested_rule.refresh_from_db()
@@ -1509,7 +1364,9 @@ def test_create_segment_obeys_max_conditions(
     # Then
     assert response.status_code == status.HTTP_400_BAD_REQUEST
     assert response.json() == {
-        "segment": "The segment has 11 conditions, which exceeds the maximum condition count of 10."
+        "segment": [
+            "The segment has 11 conditions, which exceeds the maximum condition count of 10."
+        ]
     }
     assert Segment.objects.count() == 0
 
@@ -1570,13 +1427,6 @@ def test_clone_endpoint_uses_segment_clone_service(
         "api-v1:projects:project-segments-clone", args=[project.id, segment.id]
     )
     new_segment_name = "cloned_segment"
-    service_mock = mocker.patch("segments.services.SegmentCloneService")
-    service_mock.return_value.clone.return_value = Segment(
-        id=1,
-        name="new_segment_name",
-        project=project,
-        version=1,
-    )
     data = {
         "name": new_segment_name,
     }