+++ frontend/common/types/requests.ts
@@ -85,7 +85,13 @@ export type RegisterRequest = {
   utm_data?: UtmsType
 }
 
-export type StageActionBody = { enabled: boolean; segment_id?: number }
+export type StageActionBody = {
+  enabled: boolean
+  segment_id?: number
+  initial_split?: number
+  increase_by?: number
+  increase_every?: string
+}
 export interface StageActionRequest {
   action_type: StageActionType | ''
   action_body: StageActionBody
@@ -203,6 +209,7 @@ export type Req = {
   getAvailablePermissions: { level: PermissionLevel }
   getTag: { id: string }
   getHealthEvents: { projectId: number | string }
+  dismissHealthEvent: { projectId: number | string; eventId: number }
   getHealthProviders: { projectId: number }
   createHealthProvider: { projectId: number; name: string }
   deleteHealthProvider: { projectId: number; name: string }